# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CollectionResponse:
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      description:
        example: A collection of my favorite movies
        type: string
      has_movie:
        example: true
        type: boolean
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      item_count:
        example: 5
        type: integer
      name:
        example: My Favorites
        type: string
      slug:
        example: my-favorites
        type: string
      type:
        example: custom
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      user_id:
        example: 660e8400-e29b-41d4-a716-446655440000
        type: string
      visibility:
        example: private
        type: string
    type: object
...
  handlers.CreateCollectionRequest:
    properties:
      description:
        example: A collection of my favorite movies
        maxLength: 500
        type: string
      name:
        example: My Favorites
        maxLength: 100
        minLength: 1
        type: string
      visibility:
        enum:
        - public
        - private
        example: private
        type: string
    required:
    - name
    type: object
...
  handlers.UpdateCollectionRequest:
    properties:
      description:
        example: Updated description
        maxLength: 500
        type: string
      name:
        example: Updated Name
        maxLength: 100
        minLength: 1
        type: string
      visibility:
        enum:
        - public
        - private
        example: public
        type: string
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@duskforge.io
    name: API Support
    url: https://github.com/duskforge/api
  description: A movie collection and review API powered by TMDB
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Duskforge API
  version: "1.0"
paths:
...
  /users/{userId}/collections:
    get:
      consumes:
      - application/json
      description: Get all collections for a user. Returns all collections if the
        requester is the owner, only public ones otherwise. When tmdb_id is provided,
        each collection includes a has_movie flag indicating whether the movie is
        in that collection. Use the type parameter to filter by collection type. Returns
        404 if the user is banned (non-admin callers). Returns 403 if there is a block
        between the authenticated user and the target user.
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: TMDB movie ID - when provided, adds has_movie flag to each collection
        in: query
        name: tmdb_id
        type: integer
      - description: Filter by collection type
        enum:
        - system
        - custom
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of collections
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.CollectionResponse'
                  type: array
              type: object
        "400":
          description: Invalid user ID or TMDB ID
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User blocked
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user's collections
      tags:
      - collections
    post:
      consumes:
      - application/json
      description: Create a new custom collection for the authenticated user
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: Collection details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCollectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Collection created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.CollectionResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Collection with this name already exists
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a collection
      tags:
      - collections
  /users/{userId}/collections/{slug}:
    delete:
      consumes:
      - application/json
      description: Delete a collection. System collections cannot be deleted.
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: Collection slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Collection deleted
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Collection not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a collection
      tags:
      - collections
    get:
      consumes:
      - application/json
      description: Get a collection by user ID and slug. Returns the collection if
        public or if the requester is the owner. Returns 404 if the collection owner
        is banned (non-admin callers). Returns 403 if there is a block between the
        authenticated user and the collection owner.
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: Collection slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Collection details
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.CollectionResponse'
              type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User blocked
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Collection not found or owner banned
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get collection by slug
      tags:
      - collections
    patch:
      consumes:
      - application/json
      description: Update a collection's name, description, or visibility. System
        collections cannot have their name changed.
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: Collection slug
        in: path
        name: slug
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated collection
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.CollectionResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Collection not found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Collection with this name already exists
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a collection
      tags:
      - collections
...
  /users/{userId}/collections/{slug}/items/{tmdbId}:
    delete:
      consumes:
      - application/json
      description: Remove a movie from a collection by TMDB ID
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: Collection slug
        in: path
        name: slug
        required: true
        type: string
      - description: TMDB movie ID
        in: path
        name: tmdbId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Item removed
        "400":
          description: Invalid user ID or TMDB ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Collection or item not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Remove item from collection
      tags:
      - collections
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"